#281 Tag ops classes/attributes in IM so they are easily identifiable by users
#281 Tag ops classes/attributes in IM so they are easily identifiable by users

Operations (ops) stewarded product classes are listed alongside the pds stewarded ones. The request is to clearly distinguish the ops products from the pds products. The solution was to modify the existing "pds3" section to focus on "ops". The "pds3" products are stewarded by "ops" so they are subsumed into the new "ops" section.

Resolves #281
